refactor(git): split runner.ts into focused command-runner modules
