+++ config/max-lines-baseline.txt
@@ -19,7 +19,6 @@ inline src/main/codex-accounts/runtime-home-service.ts
 inline src/main/codex-accounts/service.ts
 inline src/main/codex/hook-service.ts
 inline src/main/daemon/daemon-init.ts
-inline src/main/git/runner.ts
 inline src/main/git/status.ts
 inline src/main/git/worktree.ts
 inline src/main/github/project-view.ts
