+++ src/main/git/command-runner/abort-error.ts
@@ -1,0 +1,5 @@
+export function createAbortError(): Error {
+  const error = new Error('The operation was aborted.')
+  error.name = 'AbortError'
+  return error
+}
