+++ src/main/git/command-runner/wsl-aware-spawn.ts
@@ -1,0 +1,26 @@
+import { spawn, type ChildProcess, type SpawnOptions } from 'node:child_process'
+import { recordSubprocessSpawn } from '../../diagnostics/main-thread-churn-probe'
+import { resolveCommand } from './wsl-command-resolution'
+
+/**
+ * Spawn any command with WSL awareness.
+ * Used for non-git binaries like `rg` that also need WSL routing.
+ */
+export function wslAwareSpawn(
+  command: string,
+  args: string[],
+  options: SpawnOptions & { cwd?: string; wslDistro?: string; useWslLoginShell?: boolean }
+): ChildProcess {
+  const { wslDistro, useWslLoginShell, ...spawnOptions } = options
+  const resolved = resolveCommand(command, args, options.cwd, wslDistro, {
+    useWslLoginShell
+  })
+  const spawnStartedAt = performance.now()
+  const child = spawn(resolved.binary, resolved.args, {
+    ...spawnOptions,
+    windowsHide: true,
+    cwd: resolved.cwd
+  })
+  recordSubprocessSpawn(resolved.binary, resolved.args, performance.now() - spawnStartedAt)
+  return child
+}
